Updated to newer Python syntax where possible
These are changes that are needed for Python 3, but which still work on Python
2.5.

The command used was:

2to3 -wn -f apply -f future -f has_key -f import -f long -f paren -f reduce -f repr -f sys_exc -f tuple_params zeroinstall 0launch 0store 0alias 0desktop 0store-secure-add 0install

+++ zeroinstall/gtkui/cache.py
@@ -49,7 +49,7 @@ def get_size(path):
 		size = os.path.getsize(man)
 		for line in file(man, 'rb'):
 			if line[:1] in "XF":
-				size += long(line.split(' ', 4)[3])
+				size += int(line.split(' ', 4)[3])
 	else:
 		size = 0
 		for root, dirs, files in os.walk(path):
